clean up. fix eslint configuration. remove disabling of its' errors
