+++ .eslintrc.js
@@ -37,7 +37,8 @@ module.exports = {
     'newline-per-chained-call': 0,
     'no-confusing-arrow': 0,
     'no-console': 0,
-    'no-unused-vars': 2,
+    'no-unused-vars': 'off',
+    '@typescript-eslint/no-unused-vars': 'error',
     'no-use-before-define': 0,
     'prefer-template': 2,
     'prefer-destructuring': 0,
