+++ src/app/common/components/CopyToClipboard/CopyToClipboard.tsx
@@ -3,7 +3,6 @@ import copy from 'clipboard-copy';
 import Tooltip, { TooltipProps } from '@material-ui/core/Tooltip';
 
 interface ChildProps {
-  // eslint-disable-next-line no-unused-vars
   copy: (content: any) => void;
   style?: any;
 }
@@ -11,7 +10,6 @@ interface ChildProps {
 interface CopyToClipboardProps {
   toolTipText?: string;
   tooltipProps?: Partial<TooltipProps>;
-  // eslint-disable-next-line no-unused-vars
   children: (props: ChildProps) => React.ReactElement<any>;
   style?: any;
 }
