+++ src/app/common/components/Incentivized/Incentivized.tsx
@@ -46,7 +46,7 @@ const Incentivized = (props: IncentivizedProps) => {
     marginBottom: -3,
   };
 
-  // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
+  // @typescript-eslint/no-unused-vars
   const isCurrentRound = (round: any): boolean => {
     return round.epoch_from <= currentEpoch && round.epoch_to >= currentEpoch;
   };
