+++ src/app/common/components/Table/StyledCell/StyledCell.ts
@@ -1,7 +1,7 @@
 import TableCell from '@material-ui/core/TableCell';
 import { createStyles, Theme, withStyles } from '@material-ui/core/styles';
 
-// eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
+// @typescript-eslint/no-unused-vars
 const StyledCell = withStyles((theme: Theme) => createStyles({
     head: {
         height: 36,
