--- src/lib/utils/EthereumChain.ts
@@ -1,9 +1,0 @@
-import { BaseChain, CHAIN, Chain } from '~lib/utils/BaseChain';
-
-export class EthereumChain extends BaseChain {
-    chain: Chain = CHAIN.ETHEREUM;
-
-    getChainPrefix(): string {
-        return '';
-    }
-}
