--- src/lib/utils/GoerliChain.ts
@@ -1,5 +1,0 @@
-import { BaseChain, CHAIN, Chain } from '~lib/utils/BaseChain';
-
-export class GoerliChain extends BaseChain {
-    chain: Chain = CHAIN.GOERLI;
-}
