--- src/lib/utils/HoleskyChain.ts
@@ -1,5 +1,0 @@
-import { BaseChain, CHAIN, Chain } from '~lib/utils/BaseChain';
-
-export class HoleskyChain extends BaseChain {
-    chain: Chain = CHAIN.HOLESKY;
-}
