+++ src/lib/utils/PraterChain.ts
@@ -1,4 +1,4 @@
-import { BaseChain, CHAIN, Chain } from '~lib/utils/BaseChain';
+import { BaseChain, CHAIN, Chain } from '~lib/utils/ChainService';
 
 export class PraterChain extends BaseChain {
     chain: Chain = CHAIN.PRATER;
