--- src/lib/utils/networkHelper.ts
@@ -1,4 +1,0 @@
-export const NETWORKS = {
-    PRATER: 'prater',
-    MAINNET: 'mainnet',
-};
